requirements.txt
kafka-python
typer
rich
lbry-rocksdb